Android is a variant of Linux, but this breaks the matching process.
In the matching process, any request with Linux as a platform will also
match a node with Android as a platform. This does not make sense since
a test for Chrome/Linux could end up running in an Appium-Android node.

For the same reasons iOS had already family set to null, therefore now
setting family to null for Android.