Rethrow exceptions when RemoteSession start fails
Currently, if a remote session fails to start, then Optional.empty() is
returned and the system behaves as expected. However, RemoteSession does
not make a call to service.stop(), and ServicedSession is unaware that
an exception has occurred, so it also doesn't make this call.

Thus, in the case that the Remote Session fails, there will can
lingering processes (ex. chromedriver) on the remote host. If running in
a high-traffic environment, then these lingering processes can
eventually starve the host of resources.

Fixes lingering processes on #4958