[grid] AnnotatedConfig should ignore default values of primitive types
For primitive types, it is impossible to determine whether
or not a value has been set to the default value (false for
booleans, 0 for numbers) or has merely been left blank. The
annotated configs must therefore support the ability to
ignore those values, otherwise they'll be "set", and that
prevents the CompoundConfig from working as expected.