Reformat all BUCK and .bzl files
By invoking the bazel `buildifier` tool on them. These
files are all written using the StarLark language used
in bazel, and so this should be a safe operation.

Future changes to build files should also be formatted
this way in order to keep consistency and minimise the
diffs submitted when working on build files.