[py] Raise error for unsupported method
The Selenium wire protocol does not define a "set window rect" method
[1]. Although similar behavior can be achieved by composing available
methods, doing so could lead to observable differences with WebDriver
(for instance, depending on the order the methods are invoked) and cause
issues when migrating.

Immediately raise an exception if the `set_window_rect` method is
invoked for a browser which uses the Selenium wire protocol. Extend the
method's documentation to include a recommendation on how to update
call sites in a forward-compatible way.

[1] https://github.com/SeleniumHQ/selenium/wiki/JsonWireProtocol