Allow .NET HttpCommandExecutor subclasses to set CommandInfoRepository
This allows subclasses of HttpCommandExecutor to properly update the
CommandInfoRepository upon creating a new session. Please note that
making the property setter protected should not be assumed as part of
the permanent API contract. It may be removed at any time, without
warning, and without following the standard project deprecation
pattern.