Updating supported .NET Framework versions
This removes support for .NET Framework 3.5 and .NET Framework 4.0.
Going forward, the minimum supported framework for the .NET language
bindings is .NET 4.5. We will produce binaries for .NET Framework 4.5,
4.6, 4.7, and .NET Standard 2.0. While it would be theoretically
possible to allow the .NET Standard binary to suffice for 4.6.1 or
above, in practice, doing so adds many additional assemblies added to
the output directory, which is a suboptimal outcome. .NET Framework
4.7.1 is the first version that supports .NET Standard 2.0 without the
need for these additional assemblies.