[grid]: Modify how we create sessions in the new grid
This change preserves the downstream dialects in the New
Session request until it hits the node. This allows the
node to respond with a correctly formatted response.
This emulates the existing "passthrough" behaviour of
Grid 3, and adds an additional capability in that the
response is sent correctly formatted for the local end.