Try and pick the default host address
Our `NetworkUtils` returns the first public address it finds
in its list of NICs. This is normally fine, but there's no
guarantee that the first address found is actually routeable
or the default address of the machine. That means that
anyone doing a look up of the address by host name may not be
able to connect to the machine.

Resolve this problem by attempting to ensure that the first
NIC we return in our list of NICs is the default one.