[tracing]: Back out OpenCensus support
It has been announced that OpenTracing and OpenCensus will be merging to
create a unified API.

https://medium.com/opentracing/merging-opentracing-and-opencensus-f0fe9c7ca6f0

This merged API has the goal of supporting both extant distributed tracing
APIs, so rather than support both, we can standardise on one. I strongly
suspect that the ultimate API will look a lot like OpenCensus, but right
now OpenTracing is more widely adopted, and so that's the underlying API
we shall use.