Moving .NET IWrapsDriver and IWrapsElement interfaces to base namespace
This commit moves the IWrapsDriver and IWrapsElement interfaces from
OpenQA.Selenium.Internal to OpenQA.Selenium. This should be a no-op for
the vast majority of users, requiring only a recompile of code, given
that both of these interfaces feature return types in the base namespace,
meaning that users likely already have "using" statements for the root
namespace in their source. If errors are encountered, changing the
namespace in the code and rebuilding should resolve the error.