Adding Bazel rules for building .NET targets
The rules added handle building the .NET assemblies for .NET
Framework 4.5, 4.6, and 4.7, as well as .NET Standard 2.0.
Additionally, the added rules will also merge the dependency
Json.NET assembly into the base WebDriver.dll assembly, and
will handle the creation of NuGet packages. All of these can
be done for both the normal and the strong-named (the so-called
"signed") case.