Allow attaching an existing Internet Explorer
This attachment can be done on new session creation where the
capability 'attachExistingBrowser' is set to 'true' in ieOptions
Then, driver looks at existing Internet Explorer windows, created
programmatically (i.e: the same way IEDrvierServer creates browser).
If none is found, behaviour is the same as if IEDriverServer cannot
create a browser.
If several Internet Explorer are found, it maps to the first one.
Drawback of this approach is that we can only have one IE at the same
time.

	modified:   iedriver/BrowserFactory.cpp
	modified:   iedriver/CommandHandlers/NewSessionCommandHandler.cpp
	modified:   iedriver/IECommandExecutor.cpp
	modified:   iedriver/IECommandExecutor.h
	modified:   iedriver/WebDriverConstants.h