Updating .NET test dependency from FryProxy to BenderProxy
The BenderProxy project is nothing more than a renamed and properly
packaged version of the FryProxy fork we were using previously.
Attempts to contact the owner of FryProxy through multiple channels
concerning incorporating the dependency removal changes we require,
so the fork has now been reversioned and published under another
name. The package has been published to NuGet, as Bazel does not
currently handle checked-in dependencies.