Removing support from IE driver for "desiredCapabilities"
The W3C WebDriver Specification does not recognize any property in the
payload for the new session command other than "capabilities". The IE
driver has long supported the presence of "desiredCapabilities" as a
fallback for older, obsolete language bindings. Now that we are in the
process of creating the Selenium 4.0 language bindings, this legacy
compatibility should no longer be necessary.

Also, this commit includes some long-overdue code formatting for the
new session command handler. These changes are cosmetic only, and
do not alter functionality.