Provide actual working information for the Java Module System
Because we need to make this stuff work with the Java
Platform Module System, and currently, we don't. The changes
here break into a few types:

* We weren't declaring that we were building multi-release
  jars properly in our manifests. This is now fixed.
* Automated tools for figuring out packages were screwing
  up with our shaded Jetty, so we built a custom module-info
  just for that.
* Our server classes didn't really have module definitions,
  and we'd set things up so that the same class could appear
  in two different JARs. This was suboptimal as it cause a
  split package, which causes modularised java to barf.
* Some of our internal classes were being used in modules
  other than the one that they were declared in. This is a
  big no-no, and is also considered suboptimal.

I'm pretty sure this isn't the end of the matter, but I
think that this will get us a lot further. The ideal goal will
be to use jlink to build Grid TNG and have that work from an
image. We'll see if we can get there. Wish us luck.