Remove eclipse config files
None of the core developers use eclipse, and we always recommend
that people use IntelliJ when working on the code. These config
files are frequently out-of-sync with the IJ equivalents, yet
no-one complains on the mailing list or the IRC channel. It's
safe to assume they're just noise and baggage we no longer need.