Exposing event for after driver service process has started in .NET
This commit adds an event that fires after a driver service process
has started and is ready to receive commands from the local end.
The DriverServiceStartedEventArgs provides properties for the process
ID, a StreamReader for stdout, and a StreamReader for stderr. Combined
with the DriverServiceStarting event, the .NET bindings now allow the
user to completely silently start a driver service executable, and
to capture all of the diagnostic output written to the console by that
driver service.