Adding support for .NET legacy log API for drivers that support them
The logging APIs (`driver.Manage().Logs`) are not part of the W3C
WebDriver Specification, and users should only use them for drivers
that support the extension end points for the methods. Drivers that
support these methods will be marked with the new ISupportsLogs
interface, and the user can use a cast (using a non-throwing operator
like the `is` or the `as` operator) to check whether the API is
supported before attempting to call the logging methods. If the
driver does not implement the interface, the logging methods will
not throw (for now), but will return empty lists instead of valid
data.