Introduce atoms and java bindings for friendly locators
This is not an exhaustive set, but enough to give
idea of what needs to be done. Provided here are
new locators (called "friendly", but we can change
that) that hook into the normal `bot.locators`
mechanism. The locators are:

 * "above"
 * "below"
 * "left"
 * "right"
 * "near"

All of these are based on location in the DOM. That
is "above" means that an element matches the
locator iff its bottom edge is above the top edge
of an element we're comparing against.

"Near" is currently set to mean "within 50 pixels
of any edge of an element". This may not be the
best measure.