Correct contentEditable code in Clear atom
The clear atom would previously leave a single whitespace when clearing
contenteditable elements due to a limitation in firefox. This CL
makes it so that the whitespace is only left for firefox browsers.

Additionally, contenteditable elements previously did not unfocus after
clearing, which does not follow the W3C spec for clearing elements.
This changelist adds unfocusing code to run after a contenteditable
element is cleared.

Lastly, additional blur events would fire if the element to be cleared
had both a value and was contenteditable (eg a text input element with
a starting value and had its contenteditable attribute set to true).
This changelist corrects this by appending the contentEditable clause
to the end of the if-else chain.