[bazel] Enable uploading to maven repositories
To upload to a repo:

```
  bazel run
    --define maven_repo=https://oss.sonatype.org/service/local/staging/deploy/maven2
    --define maven_user=username
    --define maven_password=t0p_s3cr3t
    //path/to:java_export
```

As suggested above, only rules that are `java_export` are
considered fair game for pushing to a maven repo. Both file
and http protocols are supported, and manual testing has
been done to verify that artifacts upload to oss.sonatype.org
correctly.

We still need to hook up gpg signing of uploaded artifacts,
but that can wait for a follow up diff.