Correcting creation of .NET Standard assemblies
In 4.0 alpha 2, the project moved to Bazel to build the .NET bindings.
There is a subtle bug in the Bazel project's rules for building .NET
Core/.NET Standard code that generates unusable assemblies. A PR has
been submitted to the .NET rules for Bazel. In the meantime, this
commit points to a fork containing the fix. Fixes issue #7436. Or
it should. We hope. If not, it'll be back to the proverbial drawing
board.