Enable READ retries for connection pool
When using keep-alive connections for the remote connection
it can sometimes occur that the connection is interrupted.

Simply reconnecting seems to be a sufficient measure to continue
the test without any apparent negative consequences.

The documented behavior of the retries parameter is that it only
retries for idempotent HTTP verbs, so I believe this may be a
reasonable approach.