Prepare infrastructure for CDP in Ruby
Adds WebSocket support to allow using CDP from Ruby bindings. The
implementation is naive: it simply processes handshake and then writes
to the connection socket and reads from it.

   driver = Selenium::WebDriver.for(:chrome)
   driver.devtools.send('Page.navigate', {url: 'https://google.com'})

There is no built-in CDP commands/types/events support. These will be
added later.

The code relies on websocket gem as it's the only dependency-free
implementation.