[grid] Build docker images for Grid components
As the readme in this diff suggests, this expects everything to be
pushed to a local container.

This is a strawman diff to help bootstrap some explorations of how to
actually build and deploy these docker images (especially as part of a
k8s deployment) It's possible we'll move the actual image creation to
the docker-selenium project at a later date. For now, this allows
rapid iteration without needing to check out multiple projects.