[tracing] Extract parent OpenTelemetry span safely
Turns out the `fields` method returns the names of fields, but it's
not mandatory that those fields are present. To cope with this, we
blindly assume that extracting the span's parent will work, and catch
the exception in the case where the parent is missing.

See: https://github.com/open-telemetry/opentelemetry-java/issues/767