Allow netty server to handle websocket connections
This change allows us to register a handler for each URL that a user
might want to use, or a general purpose "any url is good" handler. The
design chosen allows us to send server-side messages without first
waiting for the client to send anything, which will be helpful when we
proxy websocket requests.

Specifically untested in this PR is how to get an `HttpHandler` to
respond to the same URL as a websocket handler. It will probably work,
but it's not something we need right now, so marking that as out of
scope for now.