Use ProxyManager instead of PoolManager in remote_connection
Currently urllib3.PoolManager is used in creating remote
connection to the grid. If a proxy is required to connect
to the grid, we have to use urllib3.ProxyManager instead
of urllib3.PoolManager.

We have to pass grid_conn_proxy_url as part of the Webdriver
constructor such that when WebDriver creates a remote_connction,
it can be passed to the RemoteConnection class and can be used
to create the connection to the grid. If proxy is passed,
ProxyManager will be used to create the connection. Otherwise
PoolManager will be used.

Fixes #8053