[bazel] Rework how we handle web tests
Until now, we've used a custom rule to handle browser-based
tests. This has been good, since it allows us to tailor our test
suites to the platform we run on, but it's had its drawbacks. The big
two problems have been:

 1. The Bazel IJ plugin doesn't detect our tests as tests, which makes
    the IDE integration less useful.
 2. The scripts we used didn't work on Windows.

This new change essentially delegates all tests down to `java_test`,
which is good, at the cost of requiring our test runner to be aware of
the fact that a test may be run with a browser that it doesn't
actually support. This leads to the incongruous situation of running
IE tests on macOS.