Adding pre- and post-build steps to Visual Studio project files
This change attempts to build non-.NET dependencies that are managed by
Bazel targets (JavaScript atoms, Java web server for test pages, etc.)
to make it easier to develop and test directly in Visual Studio. For
building releaseable artifacts, the command-line Bazel build should
still be used, but this will hopefully lower the barrier to entry for
people new to building the .NET bindings.