[grid] Allow cli flags to be derived dynamically
It's possible to extend the behaviour of Grid at runtime by using the
`--ext` flag. Doing so should allow new functionality to register new
flags that can be used on the command line to configure the
system. This patch introduces this capability.

Effectively, each command indcates which roles in the Grid they
fulfill (eg. "hub" is not only a router, but also a distributor,
session map, and message bus) and each set of discoverable flags
indicate which roles they apply to (typically there's only one, but
it's possible to apply in many different cases).

In addition to the discoverable flags, it is possible for each cli
command to specify additional flags, and we use this for setting the
httpd up for each of those that offer network services.