try to fix infinite loop issue
I faced up with a problem related to an infinite loop in this code snippet. I use Serenity for autotests, in turn, the Serenity uses Selenium. After a test fell, code is continuing to run. In my case the code snippet '((WrapsElement) arg).getWrappedElement()' returns Object WebElementFacadeStub type every stage of the loop.