Updating .NET test rules to work with Bazel on non-Windows OSes
The test rules for .NET should now correctly run when executed using
`bazel test`. On non-Windows OSes, this will use .NET Core by default,
but if one explicitly wants to use the .NET Framework, it will use
Mono to execute the tests (Mono must be installed for this to work).

To test using the .NET Framework, add `--define=framework=netfx` to
your Bazel command line for executing the tests.