More cleanup of .NET bindings project files
These changes should allow the .NET bindings to go from a clean
directory to a fully built and configured .NET bindings with
properly embedded resources, providing that the machine is properly
configured to build the project from the command line with Bazel.
This eliminates a barrier to entry with respect to using a .NET
IDE (Visual Studio and Visual Studio for Mac) to build the bindings
without taking some manual steps first (like building the JavaScript
automation atoms and the web server for running tests against).