[PY] Use ProxyManager instead of PoolManager in remote_connection (#8297)
* Use ProxyManager instead of PoolManager in remote_connection

Currently urllib3.PoolManager is used in creating remote
connection to the grid. If a proxy is required to connect
to the grid, we have to use urllib3.ProxyManager instead
of urllib3.PoolManager.

The proxy settings will be taken from https_proxy or http_proxy
environment variables based on the remote grid url passed to the
RemoteConnection class. If proxy environment variable is set,
ProxyManager will be used to create the connection. Otherwise
PoolManager will be used.

Fixes #8053

* Renamed the function to fix flake8 errors

Co-authored-by: Vishnuprakash Puthiya Kovilakath <610731+pkvprakash@users.noreply.github.com>
Co-authored-by: David Burns <david.burns@theautomatedtester.co.uk>