Clean up the Augmenter
The Augmenter provides a mechanism to enhance classes dynamically at
runtime, but the existing implementation had a number of problems:

 1. It wasn't clear which version people should use.
 2. Neither version properly handled "reflexive" calls, which made
    implementing new locating mechanisms harder.

This diff cleans that up. It does so by removing the JRE Proxy-based
implementation (sadly --- I'd much rather we didn't have the bytebuddy
dependency as it's very heavy) and cleaning out the implementation.

The key interfaces and methods were marked as `Beta` or otherwise had
notices that they were liable to change without notice, so although
this isn't strictly in the spirit of the rest of the Se4 release,
it'll get us over the hump.