Python remote webdriver: make default command executor match docs
The command_executor argument is used to define a URL of a remote server
or a custom remote connection. Most commonly, this is provided by the
caller, but it comes also with a reasonable default value according to
the docs: http://127.0.0.1:4444/wd/hub. The default value in the code
however does in fact not include the /wd/hub suffix, which is fixed by
this commit.