Attach RemoteWebDriver to existing session.
Adds a new constructor for RemoteWebDriver that will not start
a new test session but instead attach to an existing one.

This will allow C# developers to write code that is ment to
interact with running selenium sessions. This will for example
enable creation of customization plugins for selenium based software.