Changed how the Brython stdlib is accessed
Before the stdlib was accessed via files in the Brython Github 
repository, but it now accessed through the brython_stdlib.js file.  
Initially, I thought this method of acessing the Brython stdlib would 
not work because for some unknown reason, when brython_stdlib,js 
accessed the Indexed Database API, it caused the following error: 
"access to the Indexed Database API is denied in this context".  
However, recently I discovered in the brython() function there is an 
option to disable access to the database (which is used to load the 
Brython stdlib) and access through some other method (which I don't 
fully understand).