Changed default brython source to the latest development version
The reasoning behind this is instead of constantly changing the URL 
everytime a newer version of Brython comes out, use a URL that will 
always point to the latest development version.  If users do not like 
this they always have the option to override the default and pass in 
whatever 
version they would like instead.