fix: do not update already set preference
Inside the python binding we are reading from an already defined js file
for the frozen config and if user is trying to set those particular keys
in firefox profile, the changes aren't getting reflected. This leads to
a very inconsistent behavior on the user end.

We should not update the default preferences in the object if it is
already defined by the user using `set_preferences` and should do a key
check before storing it in the dict.