30s timeout is too short [skip ci]
A test can last 1 minute and if many tests are in the
queue, this could easily cause the timeout. Increased
to 300s (5min)

Also moved the method signature to Duration. The old
Grid had tons of problems with time units, milliseconds
in some places, seconds in other ones. Duration helps to
mitigate this.