[grid] Improving check for started containers
The previous assumption was that all containers
were reachable at `localhost`, however this changes
when the jar is executed inside a docker container.

The approach we will use is to ask users to provide
the docker host url to interact with docker, and not
mounting the `docker.sock`. When we know the docker
host, we assume the container is running there, and
therefore we can check if the container is ready that
way.