[grid] Securing 0mq when user provides a certificate
Fixes #7856

The initial idea was to use `Curve` security and reuse
the certificate(s) provided by the user when they want
HTTPS in the Grid. However, the docs for 0mq and Curve
with XPUB/XSUB and security are almost non-existent.
Therefore, the approach is to create a small but hard
to guess registration secret based on the provided
certificate, given that the user provides did not
provide a registration secret.