Refactor Bazel targets for .NET build
This commit updates the .NET build targets to generate the DevTools
Protocol (CDP) .NET bindings. Storage of the generation settings files
underneath the DevTools directory was not able to work properly with
pathing and inclusion of source files. Additionally, it also adds the
ability to push to the NuGet repository from the Bazel command line.