Allow user to specify DevTools Protocol verison in .NET
This change introduces an overload to the `CreateDevToolsSession` method
that will allow a user to request a specific version of the protocol to
use instead of automatically detecting the version used by the browser
being automated. While using this overload can in theory help a user
avoid code changes with every browser update when their code relies on
the concrete, browser-specific protocol domains, using this overload is
still fraught with peril, and should be avoided in favor of the
version-independent API the Selenium library provides.

This change also cleans up some API details, removing, among other
things, the `Start` method of the DevToolsSession object from the
public-facing API. When the `CreateDevToolsSession` method is called,
there is no need to start the session, as it is already started.