[grid] Container existence won't be checked.
Checking for container existence does one more
request to the docker daemon, and in moments of
high usage, this can query the daemon for all
containers unnecessarily. Instead of that, we
handle the response from the docker endpoint
and we log it. Ideally, we are stopping a
container when the user said `driver.quit()`,
so for the user does not really matter if the
container died or not. Logs are enough to
investigate issues.