Cache IWebElement before checking TagName is correct
When creating a SelectElement we must first check the WrappedElement tag is select and throw a UnexpectedTagNameException otherwise.

Previously the WrappedElement was not cached so WebDriver API could be called up to three times: first to check the tag is not null, second to check it's value is select (ignoring case) and a third time as a parameter of UnexpectedTagNameException if any of the previous conditions failed.

Fixes #8899