Add network interception to .NET
This commit adds the ability of a user to intercept and manipulate
network traffic between the browser and the web site being automated. To
use this feature at present, the browser being automated must support
the Chromium DevTools Protocol, and the driver instance must implement
the IDevTools interface. Attempting to use these features will cause an
exception to be thrown at runtime if the driver is not capable of this
communication. To access this feature, use the following code:

    // Assume "driver" is a valid IWebDriver instance
    INetwork network = driver.Manage().Network;

    // Set up monitoring information by adding handlers
    // and subscribing to events.

    // Begin monitoring by starting the monitor.
    await network.StartMonitoring();

With this new functionality, users can do the following:
* Monitor HTTP requests and responses between browser and web server
* Add handlers to modify requests that meet user-defined criteria
* Add handlers to supply responses for requests that meet user-defined
  criteria
* Add handlers to supply authentication credentials for sites using
  basic and digest authentication