Rework how By works
With this change, each WebDriver implementation needs to take
responsibility for handling locators. In the case of the classes that
extend `RemoteWebDriver` (which is most of them), we first check to
see if a locator directly supports remote location. If it does, we try
to find the element that way first. If that fails in a way that
indicates that the remote end didn't understand the search mechanism,
we fall back to passing the search context to the locator.

In order to be as efficient as possible, drivers store a map of which
search strategy should work.