[cdp] Use a flat namespace for selenium options
This means that rather than having `se:options = {cdp: uri}` we now
just have `se:cdp = uri`. Other capabilities that were stored in
`se:options` have also been extracted and placed into a flat
namespace.

This change makes it easier to extract the information we need
(trivially: we remove one lookup in a dictionary), and means that we
can now easily merge capabilities with these options set without
needing to do any additional work. Previously, we'd need to be aware
of `se:options` and do "deep merging", which is easy to overlook.