[grid] Setting Netty client pool size to 1
The AsyncHttpClient has built in a thread pool
to handle connections and the pool was growing
based on the way we use the client. As a result,
we were having many threads that kept objects in
memory that were not getting released, a memory
leak.

This is not an issue in the AsyncHttpClient but
most likely in the way we use the client.