[grid] Getting the Grid status from the Model
To get the status we were usually querying via
HTTP all nodes and then rendering that status
update. This could lead to a more precise status
but it will flood the Grid with HTTP requests
everytime the status endpoint is queried (which
some users do to monitor the Grid).

On the other hand, we keep the Node
status updated with a small delta
from the "real time" status. This
commit leverages that update process
and uses the data collected there
to show it to the user as the
overall Grid status.