[grid] Generalizing platform for browser driver
We handle different platform names and some of
them are not always properly matched by the
browser drivers. This change removes the specific
platform name and sends an empty field to let
the browser driver match without issues.

This is safe to do because the platform
matching has been done previously through
the Distributor and the Node matcher.

Fixes #9318