Make `*Options` mirror `Capabilities` far better
Prior to this change, although all the `*Options` classes implemented
`Capabilities`, they would not actually adhere to the
contract. Notably, this meant that browser-specific capabilities,
managed by the options class, would not be returned from
`getCapabilityNames`, and calling `getCapability` on those
browser-specific capabilities wouldn't return anything.

We had originally worked around this by modifying the hashCode of
options as part of the `hashCode` method. Now, we properly return all
the capability names, and therefore all the methods that depend upon
that (including `toString`, `hashCode`, `equals`, and `asMap`) now
operate equivalently.