[dotnet] Allow network interception to handle duplicate response headers
Headers in HTTP responses may be repeated. The previous implementation
of CDP network response handling did not take this into account. Now, if
a response has duplicate headers, the values will be combined into a
single header with the correct name, and a comma-separated list for the
value. This is complaint with RFC2616. There is an exception for the
Set-Cookie header, however, which is allowed to occur multiple times,
but is not combinable as to the rules of the aforementioned RFC. To
account for this, a new property on HttpResponseData has been added to
list the cookie headers (the `CookieHeaders` property) as found in the
response.