Start new sessions from the distributor in background threads
Prior to this change, the distributor would only be creating sessions
one at a time. This may be okay on a single server, but doesn't really
scale properly when there's a larger (distributed) Grid. We now grab
up to "the number of processors on the distributor's server" entries
from the new session queue and attempt to start them in parallel.