[grid] Using session request timeout in the RouterServer
When a new session request comes, the Router forwards it
to the SessionQueue, and the new session request will stay
in the queue up to 300 seconds (default). However, the read
timeout in the http client is 180 seconds (default), so
the request will timeout in the Router first. Which is why
we need to use the configured new session request timeout.

Fixes https://github.com/SeleniumHQ/docker-selenium/issues/1377