[dotnet] Make CDP sessions autodetect close of initial attached target
Previously, if the initial target to which a CDP session is attached is
closed, the session cannot continue to be used. This change detects that
scenario now, and will automatically reattach to another target, without
invalidating the object used by .NET code.

Additionally, this commit replaces the `ResetDevToolsSession()` method
in the `IDevTools` interface with a method called
`CloseDevToolsSession()`. The purpose of this method is to allow the
user to close a session without calling `Dispose()` directly on the
session object, but being unable to set the reference to the session
object to null in the driver instance.