[dotnet] Explicitly remove Expect header
When using the HttpClient class, the behavior for an HTTP POST request
is different between .NET Framework and .NET Core. In the .NET Framework
case, the client automatically adds an `Expect: 100-Continue` header.
When the remote end does not send a 100 response, the client hangs
waiting for it. This change explicitly removes that header for all HTTP
requests. Fixes issue #9277.