Add a --hub flag for the Node
If set, this will be used to derive the `--grid-url` and the
`--publish-events` and `--subscribe-events` values by the
`NodeOptions` and `ZeroMqEventBus` if those flags aren't also set.

This should be enough for most users to start a node with:

`./selenium node --hub 192.168.0.21`

and have what they expect to happen actually happen. Which is nice.