Fix the completion command's zsh completions
It's now possible to do `source <(selenium completion zsh)` in zsh and
have that Do The Right Thing. You can also copy the output to a
`_selenium` completion file and have zsh load that properly to avoid
executing a command every time a shell starts.

We don't currently properly handle sub-sub commands (eg. `selenium
info tab-tab` won't produce any useful results), but this is a step in
the right direction.