Make `SearchContext.findElements` have a generic return type
Why do this? The main reason is to support frameworks that wish to
extend Selenium, but return their own sub-types of `WebElement` from
`findBy`.

`findElement` can already be overridden using co-variant return types,
but the `List<WebElement>` returned by `findElements` could not be
overridden as easily. This signature change allows that.

With this change in place, subclasses of WebDriver can now do
something like this, without there being compilation issues:

```
public interface Special extends WebDriver {
    SpecialElement findElement(By locator);
    List<Specialelement> findElements(By locator);
}
```

Compiling the entire java code base suggests that this is a safe
change to make -- no tests or other calls sites needed to be modified
in order for this change to work.