[grid] Forwarding CDP connections to the right place
When having a Dynamic Grid, and a user who wants to do
CDP over Grid, we have the situation that the browser
is not running on the same host where the Node receives
the original CDP request. In this case, we are having a
Node running on one host, and the browser in a different
host/container.

For this, a forward url is being added to identify this
situation and redirect the request to the place where the
browser is running. All this because it is not possible
to access to the browser's debugger address remotely.

Fixes #9803