[java] Fix FluentWait's sample usage wrt durations
Commit 502301edeb4c4caf27aad125e8e75da3f22691f1 removed some
deprecated methods which used the deprecated TimeUnit class and kept
only the overloaded methods that use Java's java.time.Duration class.

However, it missed fixing the sample usage at the beginning of the
javadoc that still referred to these methods.
This patch fixes the issue and replaces the method calls in the sample
usage with the newer methods that accept a Duration argument.

Fixes #10207