[grid] Getting session requests from the queue ASAP
In the initial design made sense to wait a bit before
retrieving sessions again from the queue, because
the requests were posted from the queue to the
distributor.

As alphas and betas evolved, we moved to a design where
the Distributor is the one querying the session queue
for pending sessions to be served. Under this design,
it does not make sense to wait before asking the queue
for new session requests. The parameter is now set to
zero.

#10242