Add support for proxy authentication to remote_connection
Currently, urllib3.ProxyManager is used when connecting to a Selenium
Server via http proxy. The proxy url is read from environment variables.

When using a proxy which requires authentication, credentials are
usually passed in the url like this:
'http://username:password@proxy.com:8080'
urllib3.ProxyManager does not support this, but instead provides a
proxy_header field which takes a basic auth header for authentication.

This commit implements support for this.
If credentials are given in the proxy url, they are seperated out.
Then, the url without credentials and a proxy authentication header,
which is created from the url credentials, are used to create the
ProxyManager instance.