[java] Retry filter not used by default
The objective of the filter was to support
retries when the Grid components fail to
communicate, such as during start time.
However, other mechanisms exist for that now.

Retries are used for now only when communicating
with the browser driver only. We will add it in
other places as feedback from the community comes
in.

Users can still enable it in their RemoteWebDriver
by creating a custom config.

Fixes #10437