New implementation of one shot nodes
Adding a config to the distributor to automatically drain a node after it has gotten it's first
session.

The existing one shot node functionality (OneShotNode) has not been maintained. This is
understandable as it's repeating instead of reusing the functionality of LocalNode but isn't used by
many.

Kind of fixes #docker-selenium/1514
Related to #9845