[grid] Setting session-timeout as read timeout for http client
In short, using the default of 180s for timeout in the http
client used in the factories is not a good idea because a
user might set a shorter session-timeout, and therefore, a
session could be stopped while a http request is being executed.

This change sets session-timeout as the timeout for the
http clients used in the driver factories, and with that,
no session should be stopped while a http request is being
executed.

Fixes #10404