Fixing options handling for print command.
Prior to this change, when executing e.g. a `RemoteWebDriver#print`
command, the print options passed to the method were passed through to
the chromedriver instance, but serialized in a wrong format, rendering
them uneffective. The issues were:

  - Options were wrapped in an `options` key when chromedriver expected
    them to be at the top-level of the parameters dict (see [here](https://github.com/chromium/chromium/blob/main/chrome/test/chromedriver/window_commands.cc#L2246))
  - Page size options were put into a `pageSize` key, when chromedriver
    expected it to be a `page` key (see [here](https://github.com/chromium/chromium/blob/main/chrome/test/chromedriver/window_commands.cc#L539))

Also, using the `PageSize` class supplied by selenium was not possible,
due to defaulting to constant page sizes, not allowing the consumer to
override them. This was addressed by adding a second constructor
allowing to specify height/width.

All tests were performed against a v96 chromedriver.